refactor(vue-phaserjs): file-organization sweep over the components, composables, hooks, plugins and store
The one loose service at the package root, the container push, joins
the other services; everything else in the unit already sits where the
rules put it.
