+++ .agents/ledgers/file-organization.md
@@ -53,7 +53,7 @@ The question is where a thing lives and whether it exists twice — one export p
 | `packages/parse-tmx`, `packages/xml2js`                                            | 2026-09-13 | barrel contents are `ctix` output — regenerate, never hand-edit                                 |
 | `packages/vue-phaserjs` — `models`                                                 | 2026-09-13 | barrel contents are `ctix` output — regenerate, never hand-edit                                 |
 | `packages/vue-phaserjs` — `util`                                                   | 2026-09-13 | `util` became `services`; the three emit type utilities are `models/emit`                       |
-| `packages/vue-phaserjs` — `components`, `composables`, `hooks`, `plugins`, `store` | —          |                                                                                                 |
+| `packages/vue-phaserjs` — `components`, `composables`, `hooks`, `plugins`, `store` | 2026-09-13 |                                                                                                 |
 | `scripts/src/dependencyGraph`                                                      | 2026-09-12 | a command keeps its entrypoint; its functions live in `services/<command>`, types in `models`   |
 | `scripts/src/oxlint`                                                               | 2026-09-12 | `.oxlintrc.json` loads a plugin by path, so the entrypoint stays; the rules move out            |
 | `scripts/src/coderabbit`                                                           | 2026-09-12 | four verbs; what two share sits at `services/coderabbit` root                                   |
