+++ packages/vue-phaserjs/src/services/pushGameObject.test.ts
@@ -1,6 +1,6 @@
 import type { GameObjects } from "phaser";
 
-import { pushGameObject } from "#src/pushGameObject";
+import { pushGameObject } from "#src/services/pushGameObject";
 import { describe, expect, test, vi } from "vitest";
 
 const createMockGameObject = (depth?: number) => ({ depth, type: "GameObject" }) as unknown as GameObjects.GameObject;
