+++ packages/vue-phaserjs/src/services/pushGameObject.ts
